Moved strength-standards-generator and template-generators to api/ folder
