+++ src/client/api/template-generators/GeneratorInterface.js
@@ -1,5 +1,5 @@
 // @flow
-import type { ScheduleCardDataType } from '../types';
+import type { ScheduleCardDataType } from '../../types';
 
 export interface GeneratorInterface {
   /**
