+++ src/client/containers/ScheduleWrapper.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { Alert as UnstyledAlert } from 'reactstrap';
 import ScheduleCard from '../components/ScheduleCard';
-import TemplateGeneratorApi from '../template-generators/TemplateGeneratorApi';
+import TemplateGeneratorApi from '../api/template-generators/TemplateGeneratorApi';
 import type { ProgramTemplateType, TemplateModificationType } from '../types';
 
 type Props = {
